refactor: add new structure of folders in components
