+++ src/components/utils/Mask/tokens.ts
@@ -1,10 +1,6 @@
 import { inube } from "@inubekit/inubekit";
 
 const mask = {
-    zIndex: {
-        frond: 99,
-        behind: -1
-    },
     background: {
         light: inube.palette.neutral.N0,
         dark: inube.palette.neutral.N900
