+++ src/design/tokens/tokensWithReference/presente/index.ts
@@ -443,6 +443,12 @@ const presente = {
       },
     },
   },
+  mask: {
+    background: {
+      light: inube.palette.neutral.N0,
+      dark: inube.palette.neutral.N900,
+    },
+  },
 };
 
 export { presente };
