+++ src/design/tokens/tokensWithReference/sistemasEnlinea/index.ts
@@ -266,6 +266,12 @@ const sistemasenlinea = {
       },
     },
   },
+  mask: {
+    background: {
+      light: inube.palette.neutral.N0,
+      dark: inube.palette.neutral.N900,
+    },
+  },
 };
 
 export { sistemasenlinea };
