+++ src/pages/loginForm/interface.tsx
@@ -1,8 +1,8 @@
 import { Divider, Grid, Stack, Button } from "@inubekit/inubekit";
 import { Links } from "@pages/link";
-import { DynamicInputs } from "@components/login/DynamicInputs";
+import { DynamicInputs } from "@components/inputs/DynamicInputs";
 import { ILoginForm } from "@ptypes/components/login/ILoginForm";
-import { Header } from "@components/login/Header";
+import { Header } from "@components/layout/Header";
 
 const LoginFormUI = (props: ILoginForm) => {
     const {
