--- src/types/ICustomTokens/index.ts
@@ -1,18 +1,0 @@
-interface ICustomTokens {
-  modal?: { 
-    backgroundColor: string;
-    borderRadius: string;
-  };
-  mask?: {
-    background?: {
-      light: string;
-      dark: string;
-    };
-    zIndex?: {
-      frond: number;
-      behind: number;
-    };
-  };
-}
-
-export type{ ICustomTokens };
