+++ src/types/components/login/IMask/index.ts
@@ -1,6 +1,6 @@
 interface IMask {
     zIndex: string;
-    backgroundColor: string;
+    isModal: boolean;
     direction?: string;
 };
